tested friend status updating and viewing endpoints, added win/loss stat endpoint
