+++ server/controllers/dev/userDev/friendRequests.controller.ts
@@ -1,4 +1,6 @@
 import { FastifyRequest, FastifyReply } from 'fastify';
+import { getAllFriends, getOnlyFriends } from '../../utils/friendUtils.controller';
+
 
 // param: receiver is the person receiving THIS friend request from user
 export const sendReqDev = async (request: FastifyRequest, reply: FastifyReply): Promise<any> => {
@@ -120,64 +122,6 @@ export const rejectReqDev = async (request: FastifyRequest, reply: FastifyReply)
 		console.error(error);
 		return reply.code(500).send({ error: "Failed to reject friend request" });
 	}
-};
-
-export async function getAllFriends(userId: number, request: FastifyRequest) {
-	const friendships = await request.server.prisma.friend.findMany({
-		where: {
-			status: { in: ['FRIENDS', 'PENDING', 'BLOCKED'] },
-			OR: [
-				{ user1Id: userId},
-				{ user2Id: userId},
-			],
-		},
-		include: {
-			user1: true, 
-			user2: true,
-		},
-	});
-	// get the friend, not the user
-	const friends = friendships.map(f => {
-		const friendUser = f.user1Id === userId ? f.user2 : f.user1;
-		return {
-			friend: {
-				username: friendUser.username,
-				status: friendUser.status,
-			  },
-			status: f.status,
-			initiator: f.user1Id, // The user who sent the request
-		};
-	});
-	return friends;
-};
-
-export async function getOnlyFriends(userId: number, request: FastifyRequest) {
-	const friendships = await request.server.prisma.friend.findMany({
-		where: {
-			status: { in: ['FRIENDS'] },
-			OR: [
-				{ user1Id: userId},
-				{ user2Id: userId},
-			],
-		},
-		include: {
-			user1: true, 
-			user2: true,
-		},
-	});
-	// get the friend, not the user
-	const friends = friendships.map(f => {
-		const friendUser = f.user1Id === userId ? f.user2 : f.user1;
-		return {
-			friend: {
-				username: friendUser.username,
-				status: friendUser.status,
-			  },
-			status: f.status,
-			initiator: f.user1Id, // The user who sent the request
-		};
-	});
-	return friends;
 };
 
 export const viewAllFriendsDev = async (request: FastifyRequest, reply: FastifyReply): Promise<any> => {
